% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPseudotime.R, R/class.R
\docType{class}
\name{bulkPseudotime-class}
\alias{bulkPseudotime-class}
\alias{bulkPseudotime}
\title{bulkPseudotime function}
\arguments{
\item{expMat}{Expression matrix(tpm/fpkm data) for bulk RNA-seq.}

\item{timePointCol}{the color for timePointCol.}

\item{timelineCol}{the color for timelineCol.}
}
\value{
bulkPseudotime object
}
\description{
bulkPseudotime function

bulkPseudotime class
}
\section{Slots}{

\describe{
\item{\code{sample_dist_plot}}{list.}

\item{\code{sample_pca_plot}}{list.}

\item{\code{pseudotime_matrix}}{data.frame.}

\item{\code{pseudotime_matrix_long}}{data.frame.}

\item{\code{pseudotime_anno}}{data.frame.}

\item{\code{ordered_matrix_list}}{list.}

\item{\code{heatmap_list}}{list.}

\item{\code{heatmap_list_group}}{list.}
}}

