% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPseudotime.R
\name{bulkPseudotime}
\alias{bulkPseudotime}
\title{bulkPseudotime function}
\usage{
bulkPseudotime(
  expMat = NULL,
  timePointCol = NULL,
  timelineCol = NULL,
  window = 100
)
}
\arguments{
\item{expMat}{Expression matrix(tpm/fpkm data) for bulk RNA-seq.}

\item{timePointCol}{the color for timePointCol.}

\item{timelineCol}{the color for timelineCol.}

\item{window}{window size to produce predicted expression, large size needs more time,
default 100.}
}
\value{
bulkPseudotime object
}
\description{
bulkPseudotime function
}
