% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{bulkPseudotimeClass-class}
\alias{bulkPseudotimeClass-class}
\alias{bulkPseudotimeClass}
\title{bulkPseudotime class}
\description{
bulkPseudotime class
}
\section{Slots}{

\describe{
\item{\code{sample_dist_plot}}{list.}

\item{\code{sample_pca_plot}}{list.}

\item{\code{pseudotime_matrix}}{data.frame.}

\item{\code{pseudotime_matrix_long}}{data.frame.}

\item{\code{pseudotime_anno}}{data.frame.}

\item{\code{ordered_matrix_list}}{list.}

\item{\code{heatmap_list}}{list.}

\item{\code{heatmap_list_group}}{list.}
}}

