% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{pseudotime_heatmap,bulkPseudotimeClass-method}
\alias{pseudotime_heatmap,bulkPseudotimeClass-method}
\title{method for pseudotime_heatmap}
\usage{
\S4method{pseudotime_heatmap}{bulkPseudotimeClass}(
  object = NULL,
  order = NULL,
  timePointCol = NULL,
  timelineCol = NULL,
  ht_col = NULL,
  heatmap_params = list(),
  markGenes = NULL,
  fontsize = 12,
  ...
)
}
\arguments{
\item{object}{bulkPseudotime.}
}
\value{
Heatmap object
}
\description{
method for pseudotime_heatmap
}
