% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{pseudotime_heatmap}
\alias{pseudotime_heatmap}
\title{pseudotime_heatmap}
\usage{
pseudotime_heatmap(
  object,
  order = NULL,
  timePointCol = NULL,
  timelineCol = NULL,
  ht_col = NULL,
  heatmap_params = list(),
  markGenes = NULL,
  fontsize = 12,
  ...
)
}
\arguments{
\item{object}{bulkPseudotime object.}

\item{order}{which order heatmap to be choosed to plot.}

\item{timePointCol}{color for timePointCol.}

\item{timelineCol}{color for timelineCol.}

\item{ht_col}{color for heatmp.}

\item{heatmap_params}{more parameters for Heatmap function.}

\item{markGenes}{the gene names to be marked in heatmap.}

\item{fontsize}{fontsize for mark genes.}

\item{...}{other args will be ignored.}
}
\description{
pseudotime_heatmap
}
