% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{pseudotime_line,bulkPseudotime-method}
\alias{pseudotime_line,bulkPseudotime-method}
\title{pseudotime_line}
\usage{
\S4method{pseudotime_line}{bulkPseudotime}(object = NULL, genes = NULL, ncol = NULL, ...)
}
\arguments{
\item{object}{bulkPseudotime.}
}
\value{
ggplot object
}
\description{
pseudotime_line
}
