% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{pseudotime_line}
\alias{pseudotime_line}
\title{pseudotime_line}
\usage{
pseudotime_line(object, genes = NULL, ncol = NULL, ...)
}
\arguments{
\item{object}{bulkPseudotime object.}

\item{genes}{gene names for lineplot.}

\item{ncol}{columns for facet plot.}

\item{...}{other args will be ignored.}
}
\value{
ggplot object.
}
\description{
pseudotime_line
}
